refactor: SessionId 상수 UserSession 으로 이동
