+++ src/main/java/techcourse/myblog/interceptor/LoginInterceptor.java
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
-import static techcourse.myblog.users.UserController.USER_SESSION;
+import static techcourse.myblog.users.UserSession.USER_SESSION;
 
 @Component
 public class LoginInterceptor extends HandlerInterceptorAdapter {
